move test to DataFrameSuite + minor ws
