+++ sql/core/src/test/scala/org/apache/spark/sql/SQLQuerySuite.scala
@@ -23,8 +23,6 @@ import java.net.{MalformedURLException, URL}
 import java.sql.Timestamp
 import java.util.concurrent.atomic.AtomicBoolean
 
-import scala.util.Random
-
 import org.apache.spark.{AccumulatorSuite, SparkException}
 import org.apache.spark.scheduler.{SparkListener, SparkListenerJobStart}
 import org.apache.spark.sql.catalyst.util.StringUtils
@@ -2757,45 +2755,6 @@ class SQLQuerySuite extends QueryTest with SharedSQLContext {
         assert(sql("desc t").select("col_name")
           .as[String].collect().mkString(",").contains("i,p,j"))
       }
-    }
-  }
-
-  test("repartitionByRange") {
-    val data1d = Random.shuffle(0.to(9))
-    val data2d = data1d.map(i => (i, data1d.size - i))
-
-    checkAnswer(
-      data1d.toDF("val").repartitionByRange(data1d.size, $"val".asc)
-        .select(spark_partition_id().as("id"), $"val"),
-      data1d.map(i => Row(i, i)))
-
-    checkAnswer(
-      data1d.toDF("val").repartitionByRange(data1d.size, $"val".desc)
-        .select(spark_partition_id().as("id"), $"val"),
-      data1d.map(i => Row(i, data1d.size - 1 - i)))
-
-    // .repartitionByRange() assumes .asc by default if no explicit sort order is specified
-    checkAnswer(
-      data2d.toDF("a", "b").repartitionByRange(data2d.size, $"a".desc, $"b")
-        .select(spark_partition_id().as("id"), $"a", $"b"),
-      data2d.toDF("a", "b").repartitionByRange(data2d.size, $"a".desc, $"b".asc)
-        .select(spark_partition_id().as("id"), $"a", $"b"))
-  }
-
-  test("repartition with SortOrder") {
-    // passing SortOrder expressions to .repartition() should result in an informative error
-
-    def checkSortOrderErrorMsg[T](data: => Dataset[T]): Unit = {
-      val exc = intercept[IllegalArgumentException](data)
-      assert(exc.getMessage.contains("repartitionByRange"))
-    }
-
-    checkSortOrderErrorMsg {
-      Seq(0).toDF("a").repartition(2, $"a".asc)
-    }
-
-    checkSortOrderErrorMsg {
-      Seq((0, 0)).toDF("a", "b").repartition(2, $"a".asc, $"b")
     }
   }
 }
